+++ lib/cocoapods/installer/installation_options.rb
@@ -190,6 +190,17 @@ module Pod
       # Whether to skip generating the `Pods.xcodeproj` and perform only dependency resolution and downloading.
       #
       option :skip_pods_project_generation, false
+
+      # Whether to download pods in parallel before beginning the installation process
+      #
+      option :parallel_pod_downloads, false
+
+      # The size of the thread pool to use when downloading pods in parallel. Only takes effect when
+      # `parallel_pod_downloads` is `true`.
+      #
+      # Default: 40
+      #
+      option(:parallel_pod_download_thread_pool_size, 40, :boolean => false)
     end
   end
 end