+++ lib/cocoapods/installer/pod_source_downloader.rb
@@ -1,13 +1,14 @@
-require 'active_support/core_ext/string/strip'
 
 module Pod
   class Installer
-    # Controller class responsible of installing the activated specifications
+    # Controller class responsible for downloading the activated specifications
     # of a single Pod.
     #
     # @note This class needs to consider all the activated specs of a Pod.
     #
-    class PodSourceInstaller
+    class PodSourceDownloader
+      UNENCRYPTED_PROTOCOLS = %w(http git).freeze
+
       # @return [Sandbox] The installation target.
       #
       attr_reader :sandbox
@@ -25,7 +26,7 @@ module Pod
       # @return [Boolean] Whether the installer is allowed to touch the cache.
       #
       attr_reader :can_cache
-      alias_method :can_cache?, :can_cache
+      alias can_cache? can_cache
 
       # Initialize a new instance
       #
@@ -47,7 +48,7 @@ module Pod
         "<#{self.class} sandbox=#{sandbox.root} pod=#{root_spec.name}"
       end
 
-      # @return [String] The name of the pod this installer is installing.
+      # @return [String] The name of the pod this downloader is downloading.
       #
       def name
         root_spec.name
@@ -57,80 +58,61 @@ module Pod
 
       public
 
-      # @!group Installation
+      # @!group Downloading
 
       # Creates the target in the Pods project and the relative support files.
       #
       # @return [void]
       #
-      def install!
-        download_source unless predownloaded? || local?
-        PodSourcePreparer.new(root_spec, root).prepare! if local?
-        sandbox.remove_local_podspec(name) unless predownloaded? || local? || external?
-      end
+      def download!
+        verify_source_is_secure(root_spec)
+        download_result = Downloader.download(download_request, root, :can_cache => can_cache?)
 
-      # Cleans the installations if appropriate.
-      #
-      # Cleaning the installation will remove any files that are not used during the build process, based on
-      # the podspec and platforms of the target that the pod is integrated into.
-      #
-      # @see {#clean_installation}
-      #
-      # @return [void]
-      #
-      def clean!
-        clean_installation unless local?
-      end
+        if (specific_source = download_result.checkout_options) && specific_source != root_spec.source
+          sandbox.store_checkout_source(root_spec.name, specific_source)
+        end
 
-      # Locks the source files if appropriate.
-      #
-      # @return [void]
-      #
-      def lock_files!(file_accessors)
-        return if local?
-        unlocked_files = source_files(file_accessors).reject { |f| (File.stat(f).mode & 0o200).zero? }
-        FileUtils.chmod('u-w', unlocked_files)
-      end
-
-      # Unlocks the source files if appropriate.
-      #
-      # @return [void]
-      #
-      def unlock_files!(file_accessors)
-        return if local?
-        FileUtils.chmod('u+w', source_files(file_accessors))
+        sandbox.store_downloaded_pod(root_spec.name)
       end
 
       #-----------------------------------------------------------------------#
 
       private
 
-      # @!group Installation Steps
+      # @!group Download Steps
 
-      # Downloads the source of the Pod.
+      # Verify the source of the spec is secure, which is used to show a warning to the user if that isn't the case
+      # This method doesn't verify all protocols, but currently only prohibits unencrypted 'http://' and 'git://''
+      # connections.
       #
       # @return [void]
       #
-      def download_source
-        unless downloaded?
-          downloader = PodSourceDownloader.new(sandbox, podfile, specs_by_platform, :can_cache => can_cache?)
-          downloader.download!
+      def verify_source_is_secure(root_spec)
+        return if root_spec.source.nil? || (root_spec.source[:http].nil? && root_spec.source[:git].nil?)
+        source = if !root_spec.source[:http].nil?
+                   URI(root_spec.source[:http].to_s)
+                 elsif !root_spec.source[:git].nil?
+                   git_source = root_spec.source[:git].to_s
+                   return unless git_source =~ /^#{URI::DEFAULT_PARSER.make_regexp}$/
+                   URI(git_source)
+                 end
+        if UNENCRYPTED_PROTOCOLS.include?(source.scheme) && source.host != 'localhost'
+          UI.warn "'#{root_spec.name}' uses the unencrypted '#{source.scheme}' protocol to transfer the Pod. " \
+                'Please be sure you\'re in a safe network with only trusted hosts. ' \
+                'Otherwise, please reach out to the library author to notify them of this security issue.'
         end
       end
 
+      def download_request
+        Downloader::Request.new(
+          :spec => root_spec,
+          :released => released?,
+        )
+      end
+
       #-----------------------------------------------------------------------#
 
       private
-
-      # Removes all the files not needed for the installation according to the
-      # specs by platform.
-      #
-      # @return [void]
-      #
-      def clean_installation
-        cleaner = Sandbox::PodDirCleaner.new(root, specs_by_platform)
-        cleaner.clean!
-      end
 
       # @!group Convenience methods.
 
@@ -153,13 +135,6 @@ module Pod
         sandbox.pod_dir(root_spec.name)
       end
 
-      # @return [Boolean] whether the source has already been downloaded prior
-      #         to the installation process.
-      #
-      def downloaded?
-        sandbox.downloaded_pods.include?(root_spec.name)
-      end
-
       # @return [Boolean] whether the source has been pre downloaded in the
       #         resolution process to retrieve its podspec.
       #
@@ -174,22 +149,8 @@ module Pod
         sandbox.local?(root_spec.name)
       end
 
-      # @return [Boolean] whether the pod uses an external source (e.g. :podspec) in the
-      #         resolution process to retrieve its podspec.
-      #
-      def external?
-        @dependencies ||= podfile.dependencies.select(&:external?).map(&:name)
-        @dependencies.include?(root_spec.name)
-      end
-
       def released?
-        !local? && !predownloaded? && sandbox.specification(root_spec.name) != root_spec
-      end
-
-      # @return [Array<Pathname>] The paths of the source files
-      #
-      def source_files(file_accessors)
-        file_accessors.flat_map(&:source_files)
+        sandbox.specification(root_spec.name) != root_spec
       end
 
       #-----------------------------------------------------------------------#